+++ Sources/RichText/Extensions/RichText+Extension.swift
