+++ Sources/RichText/Models/RichTextEnums.swift
